requirements.txt
pytictoc~=1.5.1
nltk~=3.5
pyenchant~=3.1.1
pandas~=1.1.4
scipy~=1.5.4
matplotlib~=3.3.3